remove foreign key and check if loan is active

+++ src/main/java/com/demo/app/domain/customer/CustomerService.java
@@ -1,6 +1,8 @@
 package com.demo.app.domain.customer;
 
 import com.demo.app.domain.ObjectNotFoundException;
+import com.demo.app.domain.BusinessValidationException;
+import com.demo.app.domain.loan.LoanService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -12,6 +14,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CustomerService {
     private final CustomerRepository customerRepository;
+    private final LoanService loanService;
 
     public List<Customer> getAllCustomers() {
         return customerRepository.getAllCustomers();
@@ -21,9 +24,20 @@ public class CustomerService {
     public Customer getCustomer(Long customerId) {
         var customer = customerRepository.getCustomer(customerId);
         if (customer == null) {
-            throw new ObjectNotFoundException(String.format("Customer with id: %d not found", customerId));
+            throwCustomerNotFound(customerId);
         }
         return customer;
+    }
+
+    public void checkCustomerExists(Long customerId) {
+        boolean exists = customerRepository.customerExists(customerId);
+        if (!exists) {
+            throwCustomerNotFound(customerId);
+        }
+    }
+
+    public static void throwCustomerNotFound(Long customerId) {
+        throw new ObjectNotFoundException(String.format("Customer with id: %d not found", customerId));
     }
 
     @Transactional
@@ -32,13 +46,12 @@ public class CustomerService {
     }
 
     @Transactional
-    public Customer createCustomer(Customer customer) {
-        return customerRepository.saveCustomer(customer);
-    }
-
-    @Transactional
     public void removeCustomer(Long customerId) {
-        getCustomer(customerId);
-        customerRepository.deleteCustomerById(customerId);
+        boolean activeLoans = loanService.customerHasAnyActiveLoan(customerId);
+        if (activeLoans) {
+            throw new BusinessValidationException("Deletion failed - customer has active loans");
+        } else {
+            customerRepository.deleteCustomerById(customerId);
+        }
     }
 }
